refactor: move test_util into test
