+++ test/test_util/fake_cache_manager.dart
@@ -40,4 +40,7 @@ class FakeCacheManager extends CacheManager with ImageCacheManager {
   }) async {
     return MemoryFileSystem().file(url);
   }
+
+  @override
+  Future<void> dispose() async {}
 }
