+++ test/view/widget/note_footer_test.dart
@@ -24,9 +24,9 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:http_mock_adapter/http_mock_adapter.dart';
 import 'package:misskey_dart/misskey_dart.dart';
 
-import '../../../test_util/create_overrides.dart';
-import '../../../test_util/dummy_me_detailed.dart';
-import '../../../test_util/dummy_note.dart';
+import '../../test_util/create_overrides.dart';
+import '../../test_util/dummy_me_detailed.dart';
+import '../../test_util/dummy_note.dart';
 
 Future<ProviderContainer> setupWidget(
   WidgetTester tester, {
