+++ test/view/widget/reaction_button_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:misskey_dart/misskey_dart.dart';
 
-import '../../../test_util/create_overrides.dart';
-import '../../../test_util/dummy_note.dart';
+import '../../test_util/create_overrides.dart';
+import '../../test_util/dummy_note.dart';
 
 Future<ProviderContainer> setupWidget(
   WidgetTester tester, {
