+++ test/view/widget/reactions_viewer_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../test_util/create_overrides.dart';
-import '../../../test_util/dummy_note.dart';
+import '../../test_util/create_overrides.dart';
+import '../../test_util/dummy_note.dart';
 
 Future<ProviderContainer> setupWidget(
   WidgetTester tester, {
